requirements.txt
asgiref==3.5.2
chardet>=5.2.0
django==3.2.24
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
drf-yasg==1.21.4
freetype-py~=2.4.0
gemmi>=0.5.8
numpy==1.26.0
networkx>=2.8.8
pycifrw>=4.4.5
pymatgen>=2024.3.1
pyopengl>=3.1.7
pyside6==6.7.2
rdkit~=2023.9.6
requests~=2.32.0
progress
setuptools

